DHSCFT-634: omit rows from spine chart table if any data is not found
